add solutions for exercises 6.16 - 6.19
6.16 Anecdotes and the backend, step3
Modify the initialization of the Redux store to happen using asynchronous action creators, which are made possible by the Redux Thunk library.

6.17 Anecdotes and the backend, step4
Modify the creation of a new anecdote to happen using asynchronous action creators, made possible by the Redux Thunk library.

6.18 Anecdotes and the backend, step5
Voting does not yet save changes to the backend. Fix the situation with the help of the Redux Thunk library.

6.19 Anecdotes and the backend, step6
Implement improved notification