add solutions for exercise 6.23 - 6.24
6.23
Implement the application's notification state management using the useReducer hook and context. The notification should tell the user when a new anecdote is created or an anecdote is voted on:

6.24
The server requires that the content of the anecdote to be added is at least 5 characters long. Implement error handling for the insertion.